Rename hub pydantic models to schemas
