+++ cli/dstack/_internal/hub/routers/artifacts.py
@@ -3,8 +3,8 @@ from typing import List
 from fastapi import APIRouter, Depends
 
 from dstack._internal.core.artifact import Artifact
-from dstack._internal.hub.models import ArtifactsList
 from dstack._internal.hub.routers.util import call_backend, get_backend, get_project
+from dstack._internal.hub.schemas import ArtifactsList
 from dstack._internal.hub.security.permissions import ProjectMember
 
 router = APIRouter(
