+++ cli/dstack/_internal/hub/routers/jobs.py
@@ -4,8 +4,8 @@ from fastapi import APIRouter, Depends
 
 from dstack._internal.core.job import Job, JobHead
 from dstack._internal.hub.db.models import User
-from dstack._internal.hub.models import JobHeadList, JobsGet, JobsList
 from dstack._internal.hub.routers.util import call_backend, get_backend, get_project
+from dstack._internal.hub.schemas import JobHeadList, JobsGet, JobsList
 from dstack._internal.hub.security.permissions import Authenticated, ProjectMember
 
 router = APIRouter(prefix="/api/project", tags=["jobs"], dependencies=[Depends(ProjectMember())])
