+++ cli/dstack/_internal/hub/routers/link.py
@@ -3,8 +3,8 @@ from fastapi.responses import PlainTextResponse
 from fastapi.security.http import HTTPAuthorizationCredentials, HTTPBearer
 
 from dstack._internal.backend.local import LocalBackend
-from dstack._internal.hub.models import StorageLink
 from dstack._internal.hub.routers.util import call_backend, get_backend, get_project
+from dstack._internal.hub.schemas import StorageLink
 from dstack._internal.hub.security.permissions import ProjectMember
 
 router = APIRouter(prefix="/api/project", tags=["link"], dependencies=[Depends(ProjectMember())])
