+++ cli/dstack/_internal/hub/routers/tags.py
@@ -4,8 +4,8 @@ from fastapi import APIRouter, Depends, HTTPException, status
 
 from dstack._internal.core.repo import RepoRef
 from dstack._internal.core.tag import TagHead
-from dstack._internal.hub.models import AddTagPath, AddTagRun
 from dstack._internal.hub.routers.util import call_backend, error_detail, get_backend, get_project
+from dstack._internal.hub.schemas import AddTagPath, AddTagRun
 from dstack._internal.hub.security.permissions import ProjectMember
 
 router = APIRouter(prefix="/api/project", tags=["tags"], dependencies=[Depends(ProjectMember())])
