+++ cli/dstack/_internal/hub/services/backends/azure/configurator.py
@@ -49,7 +49,7 @@ from dstack._internal.backend.azure import AzureBackend
 from dstack._internal.backend.azure import utils as azure_utils
 from dstack._internal.backend.azure.config import AzureConfig
 from dstack._internal.hub.db.models import Project
-from dstack._internal.hub.models import (
+from dstack._internal.hub.schemas import (
     AzureProjectConfig,
     AzureProjectConfigWithCreds,
     AzureProjectConfigWithCredsPartial,
